requirements.txt
asgiref==3.8.1
decorator==5.2.1
Django==4.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.2.6
pillow==11.2.1
proglog==0.1.12
pypandoc==1.15
python-dotenv==1.1.0
sqlparse==0.5.3
tqdm==4.67.1
